chore(encryption): move encryption tests to their own folder
