+++ test/functional/client_side_encryption/spec_tests.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const path = require('path');
-const TestRunnerContext = require('./spec-runner').TestRunnerContext;
-const gatherTestSuites = require('./spec-runner').gatherTestSuites;
-const generateTopologyTests = require('./spec-runner').generateTopologyTests;
+const TestRunnerContext = require('../spec-runner').TestRunnerContext;
+const gatherTestSuites = require('../spec-runner').gatherTestSuites;
+const generateTopologyTests = require('../spec-runner').generateTopologyTests;
 
 const missingAwsConfiguration =
   process.env.AWS_ACCESS_KEY_ID == null || process.env.AWS_SECRET_ACCESS_KEY == null;
@@ -16,7 +16,7 @@ describe('Client Side Encryption', function() {
   }
 
   const testContext = new TestRunnerContext();
-  const testSuites = gatherTestSuites(path.join(__dirname, 'spec', 'client-side-encryption'));
+  const testSuites = gatherTestSuites(path.join(__dirname, '..', 'spec', 'client-side-encryption'));
   after(() => testContext.teardown());
   before(function() {
     return testContext.setup(this.configuration);
